GUACAMOLE-5: Extract common base class for AuthenticatedUser.
