+++ extensions/guacamole-auth-jdbc/modules/guacamole-auth-jdbc-base/src/main/java/org/apache/guacamole/auth/jdbc/user/AuthenticatedUser.java
@@ -22,9 +22,6 @@ package org.apache.guacamole.auth.jdbc.user;
 import java.util.Collections;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import javax.servlet.http.HttpServletRequest;
 import org.apache.guacamole.net.auth.AuthenticationProvider;
 import org.apache.guacamole.net.auth.Credentials;
 
@@ -33,48 +30,12 @@ import org.apache.guacamole.net.auth.Credentials;
  *
  * @author Michael Jumper 
  */
-public class AuthenticatedUser implements org.apache.guacamole.net.auth.AuthenticatedUser {
+public class AuthenticatedUser extends RemoteAuthenticatedUser {
 
     /**
      * The user that authenticated.
      */
     private final ModeledUser user;
-
-    /**
-     * The credentials given when this user authenticated.
-     */
-    private final Credentials credentials;
-
-    /**
-     * The AuthenticationProvider that authenticated this user.
-     */
-    private final AuthenticationProvider authenticationProvider;
-
-    /**
-     * The host from which this user authenticated.
-     */
-    private final String remoteHost;
-
-    /**
-     * Regular expression which matches any IPv4 address.
-     */
-    private static final String IPV4_ADDRESS_REGEX = "([0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3})";
-
-    /**
-     * Regular expression which matches any IPv6 address.
-     */
-    private static final String IPV6_ADDRESS_REGEX = "([0-9a-fA-F]*(:[0-9a-fA-F]*){0,7})";
-
-    /**
-     * Regular expression which matches any IP address, regardless of version.
-     */
-    private static final String IP_ADDRESS_REGEX = "(" + IPV4_ADDRESS_REGEX + "|" + IPV6_ADDRESS_REGEX + ")";
-
-    /**
-     * Pattern which matches valid values of the de-facto standard
-     * "X-Forwarded-For" header.
-     */
-    private static final Pattern X_FORWARDED_FOR = Pattern.compile("^" + IP_ADDRESS_REGEX + "(, " + IP_ADDRESS_REGEX + ")*$");
 
     /**
      * The connections which have been committed for use by this user in the
@@ -87,38 +48,6 @@ public class AuthenticatedUser implements org.apache.guacamole.net.auth.Authenti
     private final Set<String> preferredConnections =
             Collections.newSetFromMap(new ConcurrentHashMap<String, Boolean>());
 
-    /**
-     * Derives the remote host of the authenticating user from the given
-     * credentials object. The remote host is derived from X-Forwarded-For
-     * in addition to the actual source IP of the request, and thus is not
-     * trusted. The derived remote host is really only useful for logging,
-     * unless the server is configured such that X-Forwarded-For is guaranteed
-     * to be trustworthy.
-     *
-     * @param credentials
-     *     The credentials to derive the remote host from.
-     *
-     * @return
-     *     The remote host from which the user with the given credentials is
-     *     authenticating.
-     */
-    private static String getRemoteHost(Credentials credentials) {
-
-        HttpServletRequest request = credentials.getRequest();
-
-        // Use X-Forwarded-For, if present and valid
-        String header = request.getHeader("X-Forwarded-For");
-        if (header != null) {
-            Matcher matcher = X_FORWARDED_FOR.matcher(header);
-            if (matcher.matches())
-                return matcher.group(1);
-        }
-
-        // If header absent or invalid, just use source IP
-        return request.getRemoteAddr();
-
-    }
-    
     /**
      * Creates a new AuthenticatedUser associating the given user with their
      * corresponding credentials.
@@ -134,10 +63,8 @@ public class AuthenticatedUser implements org.apache.guacamole.net.auth.Authenti
      */
     public AuthenticatedUser(AuthenticationProvider authenticationProvider,
             ModeledUser user, Credentials credentials) {
-        this.authenticationProvider = authenticationProvider;
+        super(authenticationProvider, credentials);
         this.user = user;
-        this.credentials = credentials;
-        this.remoteHost = getRemoteHost(credentials);
     }
 
     /**
@@ -148,27 +75,6 @@ public class AuthenticatedUser implements org.apache.guacamole.net.auth.Authenti
      */
     public ModeledUser getUser() {
         return user;
-    }
-
-    /**
-     * Returns the credentials given during authentication by this user.
-     *
-     * @return 
-     *     The credentials given during authentication by this user.
-     */
-    @Override
-    public Credentials getCredentials() {
-        return credentials;
-    }
-
-    /**
-     * Returns the host from which this user authenticated.
-     *
-     * @return
-     *     The host from which this user authenticated.
-     */
-    public String getRemoteHost() {
-        return remoteHost;
     }
 
     /**
@@ -199,11 +105,6 @@ public class AuthenticatedUser implements org.apache.guacamole.net.auth.Authenti
      */
     public void preferConnection(String identifier) {
         preferredConnections.add(identifier);
-    }
-
-    @Override
-    public AuthenticationProvider getAuthenticationProvider() {
-        return authenticationProvider;
     }
 
     @Override
